TGBOT-101 Refactoring
